Cleaned price data repository and service
