+++ src/main/java/com/RecipeFinder/backend/controllers/PriceDataController.java
@@ -1,10 +1,13 @@
 package com.RecipeFinder.backend.controllers;
 
+import com.RecipeFinder.backend.models.PriceData;
 import com.RecipeFinder.backend.services.PriceDataService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/price")
@@ -13,9 +16,11 @@ public class PriceDataController {
     @Autowired
     private PriceDataService priceDataService;
 
-    @GetMapping
-    public void fetchAndPrintData() {
-        // This will call the service method to fetch and print data
-        priceDataService.fetchAndPrintPriceData();
+    @GetMapping("/all")
+    public List<PriceData> getAllFoodPriceData() {
+        return priceDataService.getAllFoodPriceData();
     }
+
+    // TODO: /filtered
+
 }
